requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
twilio
djangorestframework-simplejwt
drf-yasg
django-phonenumber-field
phonenumbers
twilio
celery==4.3
redis